refactor(capture): fold 14 declaration files into doc.go/types.go/aliases.go/constants.go
Un-split, not split. Commit e27769b6 minted twelve `*-types.go` files in one go, so
"the filenames agree with the history" here is one fact counted twice. The AST graph
shows no boundary under them: every one is a leaf with zero outgoing edges, and
type-aliases.go actually depends *upward* on accessor_counts_health.go (Snapshot =
CaptureSnapshot), which a "types layer" cannot do.

Merged away: accessor.go and types.go (both comment-only stubs whose layout notes were
already stale), websocket-types.go, network-types.go, enhanced-actions-types.go,
extension-logging-types.go, security-types.go, internal-types.go, buffer-types.go,
interfaces.go, type-aliases.go, queries_compat.go, recording_compat.go.

The result splits on a boundary that is real rather than nominal:
  - aliases.go — everything re-exported from another package (internal/types wire types,
    queries/circuit/recording/wsconn subsystems, compat constants and helpers)
  - types.go   — the types capture actually declares itself
  - doc.go     — package doc, now carrying an accurate file-layout map

Every declaration is preserved verbatim, including struct tags. go doc -all diff over the
exported symbol set shows only the A11yCache/SessionTracker removals from the previous
commit and the WSConnectionTracker alias form.

internal/capture: 64 -> 53 non-test files.
